Build JavaScript with esbuild
This unlocks a lot of good stuff, including:

- bundling
- minifying
- namespaces
- tree shaking
- better integrations with eslint
- TypeScript

Test plan: tried the reader, proofer, and dictionary tool on the
devserver.
