+++ Makefile
@@ -40,9 +40,9 @@ db_seed_all:
 # Development commands
 # ===============================================
 
-# Run the devserver, and live reload our CSS.
+# Run the devserver, and live reload our CSS and JS.
 devserver:
-	npx concurrently "flask run" "make tailwind_watcher"
+	npx concurrently "flask run" "make tailwind_watcher" "npm run build-js-dev"
 
 
 # Run Tailwind to build our CSS, and rebuild our CSS every time a relevant file
