+++ ambuda/static/js/hamburger-button.js
@@ -1,0 +1,15 @@
+import { $ } from './core';
+
+export default (() => {
+  function init() {
+    const $ham = $('#hamburger');
+    if ($ham) {
+      $ham.addEventListener('click', (e) => {
+        e.preventDefault();
+        $('#navbar').classList.toggle('hidden');
+      });
+    }
+  }
+
+  return { init };
+})();
