+++ ambuda/templates/proofing/pages/edit.html
@@ -81,10 +81,6 @@
 
 
 {% block main %}
-{# Load proofreading.js before Alpine. #}
-<script defer src="/static/js/proofreading.js?v=5"></script>
-<script src="https://unpkg.com/alpinejs@3.10.3/dist/cdn.min.js" defer></script>
-
 <article class="p-4">
   {% if not current_user.is_authenticated %}
   {% call components.p_warning() %}
@@ -103,7 +99,7 @@
   <pre class="p-2 bg-slate-100 my-4">{{ conflict.content }}</pre>
   {% endif %}
 
-  <form x-data="editor" x-init="init" method="POST">
+  <form x-data="proofer" method="POST">
     {{ form.csrf_token }}
     {{ form.version }}
 
@@ -129,7 +125,8 @@
         </div>
         <textarea id="content" name="content" required=""
           class="grow p-2 md:p-4 w-full resize-none"
-          :style="`font-size: ${textZoom}rem`">{{ form.content.data or "" }}</textarea>
+          :style="`font-size: ${textZoom}rem`"
+          @change="hasUnsavedChanges = true">{{ form.content.data or "" }}</textarea>
       </div>
 
       {# Raw image. #}
@@ -158,7 +155,10 @@
       <p class="my-4 text-sm">By saving your changes, you agree to release your
       contribution under the <a class="underline" href="{{ cc0 }}">CC0 (public
       domain) license</a>.</p>
-      <input class="btn btn-submit" type="submit" value="Publish changes">
+      <input class="btn btn-submit"
+             type="submit"
+             value="Publish changes"
+             @click="hasUnsavedChanges = false">
     </div>
   </form>
