+++ package.json
@@ -1,5 +1,10 @@
 {
+  "scripts": {
+    "build-js-prod": "esbuild ambuda/static/js/main.js --outfile=ambuda/static/gen/main.js --bundle --minify",
+    "build-js-dev": "esbuild ambuda/static/js/main.js --outfile=ambuda/static/gen/main.js --bundle --watch"
+  },
   "dependencies": {
+    "esbuild": "^0.15.5",
     "postcss": "^8.4.14",
     "tailwindcss": "^3.0.24"
   },
